+++ one/tests/test_one.py
@@ -447,6 +447,16 @@ class TestONECache(unittest.TestCase):
         # Attempt the same with the eid index missing
         datasets = datasets.droplevel(0).drop('session_path', axis=1)
         self.assertEqual(files, self.one._check_filesystem(datasets))
+        # Test with uuid_filenames as True
+        self.one.uuid_filenames = True
+        try:
+            for file, (uuid, _) in zip(files, datasets.iterrows()):
+                file.rename(file.with_suffix(f'.{uuid}{file.suffix}'))
+            files = self.one._check_filesystem(datasets)
+            self.assertTrue(all(files))
+            self.assertIn(datasets.index[0], files[0].name)
+        finally:
+            self.one.uuid_filenames = False
 
     def test_load_dataset(self):
         """Test One.load_dataset"""
@@ -735,6 +745,17 @@ class TestONECache(unittest.TestCase):
             raw_modified = One(cache_dir=tdir)._cache['_meta']['raw']['datasets']['date_modified']
             expected = self.one._cache['_meta']['modified_time'].strftime('%Y-%m-%d %H:%M')
             self.assertEqual(raw_modified, expected)
+            # Test file lock
+            t_now = time.time()
+            with mock.patch('one.api.time') as time_mock:
+                lock_file = Path(self.one.cache_dir).joinpath('.cache.lock')
+                lock_file.touch()
+                # We expect the process to sleep at first, then after skipping time,
+                # the stale lock file should be removed.
+                time_mock.time.side_effect = (t_now, t_now + 30)
+                self.one._save_cache(save_dir=tdir, force=True)  # force flag ignores modified time
+            self.assertFalse(lock_file.exists(), 'failed to remove stale lock file')
+            time_mock.sleep.assert_called()
 
     def test_update_cache_from_records(self):
         """Test One._update_cache_from_records"""